Move code around and add new namespace sketches
